+++ TodoQuest/TodoQuest/Source/View/Calendar/CalendarView.swift
@@ -14,24 +14,6 @@ import RxCocoa
 
 final class CalendarView: FSCalendar {
     
-    // MARK: - Reactive Properties
-    
-    fileprivate let selectedDateEvent = PublishRelay<Date>()
-    
-    // MARK: - Properties
-    
-    /// 이벤트가 새롭게 설정되면, 이전 이벤트와 비교하여 새로운 값만 이벤트 reload
-    var events: [Date] = [] {
-        didSet {
-            guard oldValue != events else { return }
-            let filteredEvent = events.filter { !oldValue.contains($0) }
-            
-            filteredEvent.forEach {
-                _ = calendar(self, numberOfEventsFor: $0)
-            }
-        }
-    }
-    
     // MARK: - Initializer
     
     override init(frame: CGRect) {
@@ -47,11 +29,10 @@ final class CalendarView: FSCalendar {
     
     // MARK: - Override Function
     
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        rowHeight = bounds.width / 7
-        weekdayHeight = rowHeight
+    override func setCurrentPage(_ currentPage: Date, animated: Bool) {
+        scrollEnabled = true
+        super.setCurrentPage(currentPage, animated: animated)
+        scrollEnabled = false
     }
     
 }
@@ -67,11 +48,8 @@ private extension CalendarView {
         locale = Locale(identifier: "ko_KR")
         formatter.timeZone = TimeZone(identifier: "Asia/Seoul")
         placeholderType = .fillHeadTail
-        scrollEnabled = true
-        headerHeight = 0
-        delegate = self
-        dataSource = self
-        
+        scrollEnabled = false
+        headerHeight = 0        
         setCurrentPage(Date(), animated: false)
         select(Date())
     }
@@ -80,7 +58,7 @@ private extension CalendarView {
         appearance.weekdayFont = .SCDream(size: 16, weight: .medium)
         appearance.weekdayTextColor = .Label.darkLabel
         
-        appearance.titleFont = .SCDream(size: 16, weight: .regular)
+        appearance.titleFont = .SCDream(size: 14, weight: .regular)
         appearance.titleDefaultColor = .CustomColors.blueGray
         appearance.titleTodayColor = .CustomColors.personal
         appearance.titlePlaceholderColor = .View.lightGrayBody
@@ -95,30 +73,6 @@ private extension CalendarView {
         appearance.eventSelectionColor = .CustomColors.personal
     }
     
-}
-
-// MARK: - CalendarView Delegate&DataSource
-
-extension CalendarView: FSCalendarDelegate, FSCalendarDataSource {
-    
-    func calendar(_ calendar: FSCalendar, numberOfEventsFor date: Date) -> Int {
-        let todayEvent = events.filter { Calendar.current.isDate($0, equalTo: date, toGranularity: .day) }
-        
-        return min(todayEvent.count, 3)
-    }
-    
-    func calendar(_ calendar: FSCalendar, didSelect date: Date, at monthPosition: FSCalendarMonthPosition) {
-        selectedDateEvent.accept(date)
-    }
-    
-}
-
-// MARK: - CalendarView Reactive Extension
-
-extension Reactive where Base: CalendarView {
-    var selectedDate: Observable<Date> {
-        base.selectedDateEvent.asObservable()
-    }
 }
 
 // MARK: - Preview
